# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ruralicos API",
    "version": "1.0.0",
    "description": "Alertas de subvenciones y ayudas rurales por WhatsApp. Generado por introspeccion con scripts/generar_openapi.js — regenerar tras cambiar rutas. Todas las rutas aceptan tambien el prefijo /v1 (alias del contrato v1)."
  },
  "servers": [
    {
      "url": "https://ruralicos-api.onrender.com/v1",
      "description": "Produccion (prefijo v1 recomendado)"
    },
    {
      "url": "https://ruralicos-api.onrender.com",
      "description": "Produccion (rutas legacy sin prefijo)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerUser": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token de usuario (POST /login-phone)"
      },
      "bearerOrg": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token de cooperativa (POST /partner/login)"
      },
      "bearerAdmin": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token de admin (POST /admin/login)"
      },
      "cronToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cron-token",
        "description": "Token de tareas internas/cron"
      }
    }
  },
  "paths": {
...
    "/admin/boletines/estado": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAdmin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/operations/alerts-quality": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAdmin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/operations/data-quality": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAdmin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/operations/scrapers-quality": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAdmin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/pipeline-runs": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAdmin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/scraper-runs": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAdmin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/tareas/pipeline-diario": {
      "post": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAdmin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/tareas/scraper": {
      "post": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAdmin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/tareas/scrapers-diario": {
      "post": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAdmin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}